tidy input

+++ input.cpp
@@ -28,37 +28,25 @@ void getInputState()
 	POINT pt;
 	GetCursorPos(&pt);
 	ScreenToClient(HWnd, &pt);
-	//short x = static_cast<short>(pt.x * baseWidth() / clientWidth());
-	//short y = static_cast<short>(pt.y * baseHeight() / clientHeight());
-	//short w = static_cast<short>(baseWidth());
-	//short h = static_cast<short>(baseHeight());
-	//if (x < 0)
-	//	x = 0;
-	//else if (x > w)
-	//	x = w;
-	//if (y < 0)
-	//	y = 0;
-	//else if (y > h)
-	//	y = h;
 	// ウィンドウサイズ変更対応
 	// baseWidth,baseHeihgt　ユーザー指定した大きさ
 	short x, y;
 	float aspect = baseWidth() / baseHeight();
 	if (clientWidth() == baseWidth() && clientHeight() == baseHeight()) {
-		x = (float)pt.x;
-		y = (float)pt.y;
+		x = (short)pt.x;
+		y = (short)pt.y;
 	}
 	else if (clientWidth() / clientHeight() >= aspect) {
 		float width_ = clientHeight() * aspect;
 		float left = (clientWidth() - width_) / 2.0f;
-		x = (pt.x - left) / clientHeight() * baseHeight();
-		y = pt.y / clientHeight() * baseHeight();
+		x = short((pt.x - left) / clientHeight() * baseHeight());
+		y = short(pt.y / clientHeight() * baseHeight());
 	}
 	else {
 		float height_ = clientWidth() / aspect;
 		float top = (clientHeight() - height_) / 2.0f;
-		x = pt.x / clientWidth() * baseWidth();
-		y = (pt.y - top) / clientWidth() * baseWidth();
+		x = short(pt.x / clientWidth() * baseWidth());
+		y = short((pt.y - top) / clientWidth() * baseWidth());
 	}
 
 	InputState[Now][MOUSE_X] = x;
